requirements.txt
streamlit
pandas
gql
anthropic
python-dotenv
requests_toolbelt
boto3
